requirements.txt
flask>=3.0.0
anthropic>=0.25.0
pdfplumber>=0.10.0
python-docx>=1.1.0
scikit-learn>=1.4.0
numpy>=1.26.0
gunicorn>=21.2.0
